[Gold IV] Title: 수들의 합 4, Time: 360 ms, Memory: 33948 KB -BaekjoonHub

+++ 백준/Gold/2015. 수들의 합 4/수들의 합 4.java
@@ -10,16 +10,16 @@ public class Main {
         StringTokenizer st=new StringTokenizer(br.readLine());
         int n=Integer.parseInt(st.nextToken());
         int k=Integer.parseInt(st.nextToken());
-        long[] sum=new long[n+1];
+        int[] sum=new int[n+1];
         long count=0;
-        Map<Long,Integer> map=new HashMap<>();
+        Map<Integer,Integer> map=new HashMap<>();
         st=new StringTokenizer(br.readLine());
         
         for(int i=1;i<=n;i++){
             sum[i]=sum[i-1]+Integer.parseInt(st.nextToken());
         }//end of input
         
-        map.put(0L,1);//처음을 위한 0
+        map.put(0,1);//처음을 위한 0
 
         for(int i=1;i<=n;i++){
             count+=map.getOrDefault(sum[i]-k,0);
